+++ src/main/kotlin/instructions/bitwise.kt
@@ -7,7 +7,7 @@ fun Instruction.shl(register1: Register, register2: Register, register3: Registe
     useful
     val r1 = register1.read()
     val r2 = register2.read()
-    val result = (r1.toInt() shl r2.toInt()).toByte()
+    val result = (r1.toInt() shl r2.toInt()).toShort()
     register3.write(result)
 }
 
@@ -15,7 +15,7 @@ fun Instruction.shr(register1: Register, register2: Register, register3: Registe
     useful
     val r1 = register1.read()
     val r2 = register2.read()
-    val result = (r1.toInt() shr r2.toInt()).toByte()
+    val result = (r1.toInt() shr r2.toInt()).toShort()
     register3.write(result)
 }
 
@@ -23,7 +23,7 @@ fun Instruction.and(register1: Register, register2: Register, register3: Registe
     useful
     val r1 = register1.read()
     val r2 = register2.read()
-    val result = (r1.toInt() and r2.toInt()).toByte()
+    val result = (r1.toInt() and r2.toInt()).toShort()
     register3.write(result)
 }
 
@@ -31,13 +31,13 @@ fun Instruction.or(register1: Register, register2: Register, register3: Register
     useful
     val r1 = register1.read()
     val r2 = register2.read()
-    val result = (r1.toInt() or r2.toInt()).toByte()
+    val result = (r1.toInt() or r2.toInt()).toShort()
     register3.write(result)
 }
 
 fun Instruction.not(register1: Register, register2: Register) {
     val r1 = register1.read()
-    val result = (r1.toInt().inv()).toByte()
+    val result = (r1.toInt().inv()).toShort()
     register2.write(result)
 }
 
@@ -45,6 +45,6 @@ fun Instruction.xor(register1: Register, register2: Register, register3: Registe
     useful
     val r1 = register1.read()
     val r2 = register2.read()
-    val result = (r1.toInt() xor r2.toInt()).toByte()
+    val result = (r1.toInt() xor r2.toInt()).toShort()
     register3.write(result)
 }
