+++ src/main/kotlin/instructions/conditions.kt
@@ -5,30 +5,30 @@ import org.cuttlefish.data.Register
 
 fun Instruction.eq(register1: Register, register2: Register, register3: Register) {
     useful
-    register3.write(if (register1.read() == register2.read()) 1 else 0)
+    register3.write(if (register1.read() == register2.read()) 1.toShort() else 0.toShort())
 }
 
 fun Instruction.neq(register1: Register, register2: Register, register3: Register) {
     useful
-    register3.write(if (register1.read() != register2.read()) 1 else 0)
+    register3.write(if (register1.read() != register2.read()) 1.toShort() else 0.toShort())
 }
 
 fun Instruction.lt(register1: Register, register2: Register, register3: Register) {
     useful
-    register3.write(if (register1.read() < register2.read()) 1 else 0)
+    register3.write(if (register1.read() < register2.read()) 1.toShort() else 0.toShort())
 }
 
 fun Instruction.gt(register1: Register, register2: Register, register3: Register) {
     useful
-    register3.write(if (register1.read() > register2.read()) 1 else 0)
+    register3.write(if (register1.read() > register2.read()) 1.toShort() else 0.toShort())
 }
 
 fun Instruction.lte(register1: Register, register2: Register, register3: Register) {
     useful
-    register3.write(if (register1.read() <= register2.read()) 1 else 0)
+    register3.write(if (register1.read() <= register2.read()) 1.toShort() else 0.toShort())
 }
 
 fun Instruction.gte(register1: Register, register2: Register, register3: Register) {
     useful
-    register3.write(if (register1.read() >= register2.read()) 1 else 0)
+    register3.write(if (register1.read() >= register2.read()) 1.toShort() else 0.toShort())
 }
